pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "integra-ai"
version = "0.1.0"
description = "Integra.AI: CLI + Web to generate API integrations using Google Gemini."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Integra.AI"}]
license = {text = "MIT"}
dependencies = [
  "typer[all]>=0.12.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.0",
  "Flask>=3.0.0",
  "pydantic>=2.4.0",
  "structlog>=24.1.0"
]

[project.scripts]
integra = "integra_ai.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["integra_ai*"]
